requirements.txt
flask==2.3.2
langchain==0.0.207
openai==0.27.8
requests==2.31.0
schedule==1.2.0
python-dotenv==1.0.0